pyproject.toml
[project]
name = "custom-news"
version = "0.1.0"
description = "Aplicación de noticias personalizadas con audio generado por IA"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "feedparser>=6.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "newspaper3k>=0.2.8",
    "lxml>=5.0.0",
    "lxml-html-clean>=0.1.0",
    "python-telegram-bot>=20.0",
    "streamlit>=1.30.0",
    "apscheduler>=3.10.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "aiosqlite>=0.19.0",
    "sqlalchemy>=2.0.0",
    "tenacity>=8.2.0",
    "httpx>=0.25.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
